bowtie now catches pipe failures.
If a corrupted FASTQ is input, the process creates corrupted output but does not record the error and cause Nextflow to realize the process has failed.  

This is because the error comes from the initial zcat and since it goes through a set of unix pipes, the exit code is not caught.  

This also changes the formatting of the .command.sh file so that each separate process in the pipe is on a separate line.